refactor: Rename and move widgets folder to lib/src/
